# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: CertifyPro API
  description: School Certificate Management Platform - REST API
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Log in and receive access + refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, description: "Email, school login_id, or distributor username depending on role" }
                password: { type: string }
                role: { type: string, enum: [superAdmin, schoolAdmin, distributor] }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      email: { type: string }
                      role: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /schools:
    get:
      summary: List all schools (Super Admin only)
      responses:
        '200':
          description: Array of schools with admin, wallet, and distributor info
    post:
      summary: Create a new school (Super Admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, adminName, adminEmail]
              properties:
                name: { type: string }
                adminName: { type: string }
                adminEmail: { type: string }
                adminMobile: { type: string }
                udise_code: { type: string }
                village: { type: string }
                city: { type: string }
                district: { type: string }
                taluka: { type: string }
                pin_code: { type: string }
                phone: { type: string }
                medium: { type: string }
                board: { type: string }
                distributorId: { type: string }
      responses:
        '201':
          description: School created, wallet initialized, setup email sent
...
  /students:
    get:
      summary: List students for the logged-in school
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: standard
          schema: { type: string }
        - in: query
          name: division
          schema: { type: string }
        - in: query
          name: gender
          schema: { type: string }
      responses:
        '200':
          description: Array of students
    post:
      summary: Add a new student (with optional photo upload)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name]
              properties:
                full_name: { type: string }
                photo: { type: string, format: binary }
      responses:
        '201':
          description: Student created
...
  /distributors:
    get:
      summary: List all distributors (Super Admin only)
      responses:
        '200': { description: Array of distributors }
    post:
      summary: Create a new distributor (Super Admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name: { type: string }
                email: { type: string }
                mobile: { type: string }
                city: { type: string }
                district: { type: string }
                commission_rate: { type: number }
      responses:
        '201': { description: Distributor created, setup email sent }
...
  /master-data:
    get:
      summary: List master data items, optionally filtered by category
      parameters:
        - in: query
          name: category
          schema: { type: string, enum: [standard, division, district, taluka, city, medium, religion, caste, grant_type, board_name, management_type] }
      responses:
        '200': { description: Master data items or grouped categories }
    post:
      summary: Add a master data item (Super Admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [category, value]
              properties:
                category: { type: string }
                value: { type: string }
      responses:
        '201': { description: Item created }
...